requirements.txt
# Google API dependencies
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Calendar generation
icalendar>=5.0.0

# Date parsing
python-dateutil>=2.8.2

# Timezone handling
pytz>=2024.1

# Configuration
pyyaml>=6.0

# Local LLM for email parsing (privacy-first, no cloud APIs)
ollama>=0.1.0
